backend/django_backend/requirements.txt
asgiref==3.7.2
Django==5.0.3
django-environ==0.11.2
djangorestframework==3.15.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.4.4
upgrade-pip==0.1.4

